+++ src/App.tsx
@@ -1,20 +1,17 @@
 import './App.css'
-import Card from './card'
+import Navbar from './components/navbar.tsx';
+import {Route, Routes} from "react-router-dom"
+import {Mentors} from "./pages/Mentors.tsx"
 
 function App() {
-  const userProfile = {
-    name: 'John Doe',
-    avatar: 'https://media.istockphoto.com/id/464988959/photo/mallard-duck-on-white-background.jpg?s=612x612&w=0&k=20&c=S1jcDuyXuoCVUaTobTrZ5f6SlscukkyheqKDHAeflW8=', // Sample avatar image URL
-  };
-  const attributes = ['BU', 'Class of 2026', 'Computer Science', 'Vietnam', 'Topics: Math']; // Sample attributes
-
   return (
     <>
       <div className="App">
-        <h1 id = "intro-container">Meet with your mentors</h1>
-        <div className = "w-1/2">
-          <Card userProfile={userProfile} attributes={attributes} />
-        </div>
+        <Navbar/>
+        <Routes>
+          <Route path="/"/>
+          <Route path="/mentors" element={<Mentors/>}/>
+        </Routes>
       </div>
     </>
   )
